Incorporate AspectModelBuilder functionality in AspectModelLoader
Creating an AspectModel from a list of AspectModelFiles should be located next
to the other methods that create an AspectModel from various input sources
